Refactors component directories for clarity
