+++ src/easydiffraction/experiments/components/experiment_type.py
