+++ src/easydiffraction/experiments/components/instrument.py
