+++ src/easydiffraction/experiments/components/peak.py
